requirements.txt
flask
flask-cors
tensorflow
opencv-python
numpy
pillow
python-dotenv
groq
gunicorn
scikit-learn